+++ ui/src/App/App.tsx
@@ -5,9 +5,10 @@ import PageNotFound from "./PageNotFound"
 import Dashboard from "../Dashboard/Dashboard"
 import Services from "../Services/Services"
 
-import './App.css'
+import './App.scss'
 import JumpToMenu from "./JumpToMenu"
 import {GlobalHotKeys} from "react-hotkeys"
+import Doc from "../Docs/Doc"
 
 const keyMap = {
     TOGGLE_JUMP_TO_MENU: ["command+k", "control+k"],
@@ -27,6 +28,8 @@ class App extends Component {
                             {/*<Route path="/tasks" component={Dashboard}/>*/}
                             {/*<Route path="/sources" component={Dashboard}/>*/}
                             {/*<Route path="/assets" component={Dashboard}/>*/}
+                            <Redirect exact from="/docs" to="/docs/readme" />
+                            <Route path="/docs/*" component={Doc} />
                             <Route component={PageNotFound}/>
                         </Switch>
                     </div>